+++ custom_components/ge_home/climate.py
@@ -1,35 +1,39 @@
 """GE Home Climate Entities"""
-import async_timeout
 import logging
 from typing import Callable
 
 from homeassistant.components.climate import ClimateEntity
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.core import HomeAssistant
+from homeassistant.core import HomeAssistant, callback
+from homeassistant.helpers.dispatcher import async_dispatcher_connect
+from homeassistant.helpers import entity_registry as er
 
 from .entities import GeClimate
 from .const import DOMAIN
+from .devices import ApplianceApi
 from .update_coordinator import GeHomeUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
 async def async_setup_entry(hass: HomeAssistant, config_entry: ConfigEntry, async_add_entities: Callable):
-    """GE Home Water Heaters."""
+    """GE Climate Devices."""
+    
     _LOGGER.debug('Adding GE Climate Entities')
     coordinator: GeHomeUpdateCoordinator = hass.data[DOMAIN][config_entry.entry_id]
+    registry = er.async_get(hass)
 
-    # This should be a NOP, but let's be safe
-    with async_timeout.timeout(20):
-        await coordinator.initialization_future
-    _LOGGER.debug('Coordinator init future finished')
+    @callback
+    def async_devices_discovered(apis: list[ApplianceApi]):
+        _LOGGER.debug(f'Found {len(apis):d} appliance APIs')
 
-    apis = list(coordinator.appliance_apis.values())
-    _LOGGER.debug(f'Found {len(apis):d} appliance APIs')
-    entities = [
-        entity 
-        for api in apis 
-        for entity in api.entities
-        if isinstance(entity, GeClimate)
-    ]
-    _LOGGER.debug(f'Found {len(entities):d} climate entities')
-    async_add_entities(entities)
+        entities = [
+            entity 
+            for api in apis 
+            for entity in api.entities
+            if isinstance(entity, GeClimate)
+            if not registry.async_is_registered(entity.entity_id)
+        ]
+        _LOGGER.debug(f'Found {len(entities):d} unregistered climate entities')
+        async_add_entities(entities)
+
+    async_dispatcher_connect(hass, coordinator.signal_ready, async_devices_discovered)