+++ custom_components/ge_home/entities/__init__.py
@@ -7,5 +7,6 @@ from .advantium import *
 from .ac import *
 from .hood import *
 from .water_softener import *
+from .water_heater import *
 from .opal_ice_maker import *
 from .ccm import *