+++ custom_components/ge_home/entities/oven/ge_oven.py
@@ -13,12 +13,12 @@ from gehomesdk import (
 from homeassistant.const import ATTR_TEMPERATURE, TEMP_CELSIUS, TEMP_FAHRENHEIT
 from ...const import DOMAIN
 from ...devices import ApplianceApi
-from ..common import GeWaterHeater
+from ..common import GeAbstractWaterHeater
 from .const import *
 
 _LOGGER = logging.getLogger(__name__)
 
-class GeOven(GeWaterHeater):
+class GeOven(GeAbstractWaterHeater):
     """GE Appliance Oven"""
 
     icon = "mdi:stove"
@@ -41,7 +41,7 @@ class GeOven(GeWaterHeater):
 
     @property
     def unique_id(self) -> str:
-        return f"{DOMAIN}_{self.serial_number}_{self.oven_select.lower()}"
+        return f"{DOMAIN}_{self.serial_or_mac}_{self.oven_select.lower()}"
 
     @property
     def name(self) -> Optional[str]:
@@ -50,7 +50,7 @@ class GeOven(GeWaterHeater):
         else:
             oven_title = "Oven"
 
-        return f"{self.serial_number} {oven_title}"
+        return f"{self.serial_or_mac} {oven_title}"
 
     @property
     def temperature_unit(self):