+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 
 # GE Home Appliances (SmartHQ) Changelog
 
+## 0.6.5
+
+- Added beverage cooler support (@kksligh)
+- Added dual dishwasher support (@jkili)
+- Added initial espresso maker support (@datagen24)
+- Added whole home water heater support (@seantibor)
+
+## 0.6.3
+
+- Updated detection of invalid serial numbers (#89)
+- Updated implementation of number entities to fix deprecation warnings (#85)
+
+## 0.6.2
+
+- Fixed issue with water heater naming when no serial is present
+- Initial support for built-in air conditioners (@DaveZheng)
+
 ## 0.6.1
 
 - Fixed issue with water filter life sensor (@rgabrielson11)